refactor: change file name from user to users #17
