+++ src/app.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { RoomsModule } from './rooms/rooms.module';
-import { UserModule } from './users/user.module';
+import { UserModule } from './users/users.module';
 
 @Module({
   imports: [RoomsModule, UserModule],
