+++ src/rooms/users/users.module.ts
@@ -3,7 +3,7 @@ import { RoomUsersService } from './users.service';
 import { RoomUsersController } from './users.controller';
 import { AuthModule } from 'src/guard/auth/auth.module';
 import { RoomsService } from '../rooms.service';
-import { UserService } from 'src/users/user.service';
+import { UserService } from 'src/users/users.service';
 
 @Module({
   controllers: [RoomUsersController],
