+++ src/users/users.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UserService } from './user.service';
+import { UserService } from './users.service';
 
 describe('UserService', () => {
   let service: UserService;
